refactor:유틸 분리
